# PocCompassFunc/openapi.json
{
 "components": {
  "schemas": {
...
   "Assumption": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "priority": {
      "$ref": "#/components/schemas/Priority",
      "default": "medium"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "status": {
      "$ref": "#/components/schemas/AssumptionStatus",
      "default": "untested"
     },
     "text": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     },
     "type": {
      "const": "assumption",
      "default": "assumption",
      "title": "Type",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "type",
     "text",
     "status",
     "priority",
     "id",
     "projectId",
     "created_at",
     "updated_at",
     "created_by"
    ],
    "title": "Assumption",
    "type": "object"
   },
   "AssumptionFields": {
    "properties": {
     "priority": {
      "$ref": "#/components/schemas/Priority",
      "default": "medium"
     },
     "status": {
      "$ref": "#/components/schemas/AssumptionStatus",
      "default": "untested"
     },
     "text": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     },
     "type": {
      "const": "assumption",
      "default": "assumption",
      "title": "Type",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "AssumptionFields",
    "type": "object"
   },
...
   "BulkTaskCreate": {
    "properties": {
     "tasks": {
      "items": {
       "$ref": "#/components/schemas/TaskFields"
      },
      "maxItems": 50,
      "minItems": 1,
      "title": "Tasks",
      "type": "array"
     }
    },
    "required": [
     "tasks"
    ],
    "title": "BulkTaskCreate",
    "type": "object"
   },
...
   "ChatApply": {
    "properties": {
     "operation_indexes": {
      "anyOf": [
       {
        "items": {
         "type": "integer"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Operation Indexes"
     }
    },
    "title": "ChatApply",
    "type": "object"
   },
   "ChatMessage": {
    "properties": {
     "content": {
      "title": "Content",
      "type": "string"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "proposal": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Proposal"
       },
       {
        "type": "null"
       }
      ]
     },
     "role": {
      "$ref": "#/components/schemas/ChatRole"
     },
     "user_email": {
      "title": "User Email",
      "type": "string"
     }
    },
    "required": [
     "id",
     "projectId",
     "user_email",
     "role",
     "content",
     "created_at",
     "proposal"
    ],
    "title": "ChatMessage",
    "type": "object"
   },
...
   "ChatSend": {
    "properties": {
     "message": {
      "minLength": 1,
      "title": "Message",
      "type": "string"
     }
    },
    "required": [
     "message"
    ],
    "title": "ChatSend",
    "type": "object"
   },
   "ChatSendResponse": {
    "properties": {
     "assistant_message": {
      "$ref": "#/components/schemas/ChatMessage"
     },
     "user_message": {
      "$ref": "#/components/schemas/ChatMessage"
     }
    },
    "required": [
     "user_message",
     "assistant_message"
    ],
    "title": "ChatSendResponse",
    "type": "object"
   },
   "Criterion": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "status": {
      "$ref": "#/components/schemas/CriterionStatus",
      "default": "not_met"
     },
     "target": {
      "default": "",
      "maxLength": 200,
      "title": "Target",
      "type": "string"
     },
     "text": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     },
     "type": {
      "const": "criterion",
      "default": "criterion",
      "title": "Type",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "type",
     "text",
     "target",
     "status",
     "id",
     "projectId",
     "created_at",
     "updated_at",
     "created_by"
    ],
    "title": "Criterion",
    "type": "object"
   },
...
   "DecisionItem": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "decision": {
      "$ref": "#/components/schemas/Decision"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "note": {
      "default": "",
      "maxLength": 1000,
      "title": "Note",
      "type": "string"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "type": {
      "const": "decision",
      "default": "decision",
      "title": "Type",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "type",
     "decision",
     "note",
     "id",
     "projectId",
     "created_at",
     "updated_at",
     "created_by"
    ],
    "title": "DecisionItem",
    "type": "object"
   },
...
   "Evidence": {
    "properties": {
     "assumption_id": {
      "minLength": 1,
      "title": "Assumption Id",
      "type": "string"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "result": {
      "$ref": "#/components/schemas/EvidenceResult"
     },
     "source": {
      "default": "",
      "maxLength": 500,
      "title": "Source",
      "type": "string"
     },
     "summary": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Summary",
      "type": "string"
     },
     "type": {
      "const": "evidence",
      "default": "evidence",
      "title": "Type",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "type",
     "assumption_id",
     "summary",
     "result",
     "source",
     "id",
     "projectId",
     "created_at",
     "updated_at",
     "created_by"
    ],
    "title": "Evidence",
    "type": "object"
   },
...
   "Feedback": {
    "properties": {
     "content_hash": {
      "title": "Content Hash",
      "type": "string"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "judgement": {
      "$ref": "#/components/schemas/FeedbackJudgement"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "task_id": {
      "title": "Task Id",
      "type": "string"
     },
     "type": {
      "const": "feedback",
      "default": "feedback",
      "title": "Type",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "type",
     "task_id",
     "content_hash",
     "judgement",
     "id",
     "projectId",
     "created_at",
     "updated_at",
     "created_by"
    ],
    "title": "Feedback",
    "type": "object"
   },
   "FeedbackCreate": {
    "properties": {
     "judgement": {
      "$ref": "#/components/schemas/FeedbackJudgement"
     },
     "task_id": {
      "title": "Task Id",
      "type": "string"
     }
    },
    "required": [
     "task_id",
     "judgement"
    ],
    "title": "FeedbackCreate",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Project": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "deadline": {
      "format": "date",
      "title": "Deadline",
      "type": "string"
     },
     "deleted": {
      "default": false,
      "title": "Deleted",
      "type": "boolean"
     },
     "goal": {
      "title": "Goal",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "members": {
      "items": {
       "type": "string"
      },
      "title": "Members",
      "type": "array"
     },
     "mode": {
      "$ref": "#/components/schemas/Mode"
     },
     "owner_email": {
      "title": "Owner Email",
      "type": "string"
     },
     "start_date": {
      "format": "date",
      "title": "Start Date",
      "type": "string"
     },
     "status": {
      "$ref": "#/components/schemas/ProjectStatus",
      "default": "active"
     },
     "summary": {
      "$ref": "#/components/schemas/ProjectSummary"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "mode",
     "title",
     "goal",
     "start_date",
     "deadline",
     "status",
     "owner_email",
     "members",
     "deleted",
     "created_at",
     "updated_at",
     "summary"
    ],
    "title": "Project",
    "type": "object"
   },
...
   "ProjectUpdate": {
    "properties": {
     "deadline": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline"
     },
     "goal": {
      "anyOf": [
       {
        "maxLength": 2000,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Goal"
     },
     "members": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Members"
     },
     "mode": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Mode"
       },
       {
        "type": "null"
       }
      ]
     },
     "start_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "status": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ProjectStatus"
       },
       {
        "type": "null"
       }
      ]
     },
     "title": {
      "anyOf": [
       {
        "maxLength": 100,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "title": "ProjectUpdate",
    "type": "object"
   },
...
   "Task": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "description": {
      "default": "",
      "maxLength": 1000,
      "title": "Description",
      "type": "string"
     },
     "due_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Due Date"
     },
     "effort_hours": {
      "anyOf": [
       {
        "exclusiveMinimum": 0.0,
        "maximum": 10000.0,
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Effort Hours"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "linked_assumption_ids": {
      "items": {
       "type": "string"
      },
      "title": "Linked Assumption Ids",
      "type": "array"
     },
     "linked_criterion_ids": {
      "items": {
       "type": "string"
      },
      "title": "Linked Criterion Ids",
      "type": "array"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "status": {
      "$ref": "#/components/schemas/TaskStatus",
      "default": "todo"
     },
     "title": {
      "maxLength": 100,
      "minLength": 1,
      "title": "Title",
      "type": "string"
     },
     "type": {
      "const": "task",
      "default": "task",
      "title": "Type",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "type",
     "title",
     "description",
     "status",
     "effort_hours",
     "due_date",
     "linked_assumption_ids",
     "linked_criterion_ids",
     "id",
     "projectId",
     "created_at",
     "updated_at",
     "created_by"
    ],
    "title": "Task",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "PoC Compass API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/projects/{project_id}": {
   "delete": {
    "operationId": "delete_project_api_projects__project_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Project",
    "tags": [
     "projects"
    ]
   },
   "get": {
    "operationId": "get_project_api_projects__project_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Project"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Project",
    "tags": [
     "projects"
    ]
   },
   "patch": {
    "operationId": "update_project_api_projects__project_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Project"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Project",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/projects/{project_id}/chat": {
   "delete": {
    "operationId": "clear_api_projects__project_id__chat_delete",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Clear",
    "tags": [
     "chat"
    ]
   },
   "get": {
    "operationId": "list_messages_api_projects__project_id__chat_get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ChatMessage"
         },
         "title": "Response List Messages Api Projects  Project Id  Chat Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Messages",
    "tags": [
     "chat"
    ]
   },
   "post": {
    "operationId": "send_api_projects__project_id__chat_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatSend"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatSendResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Send",
    "tags": [
     "chat"
    ]
   }
  },
  "/api/projects/{project_id}/chat/{message_id}/apply": {
   "post": {
    "operationId": "apply_api_projects__project_id__chat__message_id__apply_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "message_id",
      "required": true,
      "schema": {
       "title": "Message Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/ChatApply"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatMessage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Apply",
    "tags": [
     "chat"
    ]
   }
  },
  "/api/projects/{project_id}/chat/{message_id}/discard": {
   "post": {
    "operationId": "discard_api_projects__project_id__chat__message_id__discard_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "message_id",
      "required": true,
      "schema": {
       "title": "Message Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatMessage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Discard",
    "tags": [
     "chat"
    ]
   }
  },
...
  "/api/projects/{project_id}/feedback": {
   "post": {
    "operationId": "add_feedback_api_projects__project_id__feedback_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FeedbackCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "discriminator": {
          "mapping": {
           "assumption": "#/components/schemas/Assumption",
           "criterion": "#/components/schemas/Criterion",
           "decision": "#/components/schemas/DecisionItem",
           "evidence": "#/components/schemas/Evidence",
           "feedback": "#/components/schemas/Feedback",
           "task": "#/components/schemas/Task"
          },
          "propertyName": "type"
         },
         "oneOf": [
          {
           "$ref": "#/components/schemas/Assumption"
          },
          {
           "$ref": "#/components/schemas/Criterion"
          },
          {
           "$ref": "#/components/schemas/Task"
          },
          {
           "$ref": "#/components/schemas/Evidence"
          },
          {
           "$ref": "#/components/schemas/DecisionItem"
          },
          {
           "$ref": "#/components/schemas/Feedback"
          }
         ],
         "title": "Response Add Feedback Api Projects  Project Id  Feedback Post"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Add Feedback",
    "tags": [
     "items"
    ]
   }
  },
  "/api/projects/{project_id}/items": {
   "get": {
    "operationId": "list_items_api_projects__project_id__items_get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectItems"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Items",
    "tags": [
     "items"
    ]
   },
   "post": {
    "operationId": "create_item_api_projects__project_id__items_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "discriminator": {
         "mapping": {
          "assumption": "#/components/schemas/AssumptionFields",
          "criterion": "#/components/schemas/CriterionFields",
          "decision": "#/components/schemas/DecisionFields",
          "evidence": "#/components/schemas/EvidenceFields",
          "task": "#/components/schemas/TaskFields"
         },
         "propertyName": "type"
        },
        "oneOf": [
         {
          "$ref": "#/components/schemas/AssumptionFields"
         },
         {
          "$ref": "#/components/schemas/CriterionFields"
         },
         {
          "$ref": "#/components/schemas/TaskFields"
         },
         {
          "$ref": "#/components/schemas/EvidenceFields"
         },
         {
          "$ref": "#/components/schemas/DecisionFields"
         }
        ],
        "title": "Body"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "discriminator": {
          "mapping": {
           "assumption": "#/components/schemas/Assumption",
           "criterion": "#/components/schemas/Criterion",
           "decision": "#/components/schemas/DecisionItem",
           "evidence": "#/components/schemas/Evidence",
           "feedback": "#/components/schemas/Feedback",
           "task": "#/components/schemas/Task"
          },
          "propertyName": "type"
         },
         "oneOf": [
          {
           "$ref": "#/components/schemas/Assumption"
          },
          {
           "$ref": "#/components/schemas/Criterion"
          },
          {
           "$ref": "#/components/schemas/Task"
          },
          {
           "$ref": "#/components/schemas/Evidence"
          },
          {
           "$ref": "#/components/schemas/DecisionItem"
          },
          {
           "$ref": "#/components/schemas/Feedback"
          }
         ],
         "title": "Response Create Item Api Projects  Project Id  Items Post"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Item",
    "tags": [
     "items"
    ]
   }
  },
  "/api/projects/{project_id}/items/bulk": {
   "post": {
    "operationId": "bulk_create_api_projects__project_id__items_bulk_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BulkTaskCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "discriminator": {
           "mapping": {
            "assumption": "#/components/schemas/Assumption",
            "criterion": "#/components/schemas/Criterion",
            "decision": "#/components/schemas/DecisionItem",
            "evidence": "#/components/schemas/Evidence",
            "feedback": "#/components/schemas/Feedback",
            "task": "#/components/schemas/Task"
           },
           "propertyName": "type"
          },
          "oneOf": [
           {
            "$ref": "#/components/schemas/Assumption"
           },
           {
            "$ref": "#/components/schemas/Criterion"
           },
           {
            "$ref": "#/components/schemas/Task"
           },
           {
            "$ref": "#/components/schemas/Evidence"
           },
           {
            "$ref": "#/components/schemas/DecisionItem"
           },
           {
            "$ref": "#/components/schemas/Feedback"
           }
          ]
         },
         "title": "Response Bulk Create Api Projects  Project Id  Items Bulk Post",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bulk Create",
    "tags": [
     "items"
    ]
   }
  },
  "/api/projects/{project_id}/items/{item_id}": {
   "delete": {
    "operationId": "delete_item_api_projects__project_id__items__item_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "title": "Item Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Item",
    "tags": [
     "items"
    ]
   },
   "patch": {
    "operationId": "update_item_api_projects__project_id__items__item_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "item_id",
      "required": true,
      "schema": {
       "title": "Item Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ItemPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "discriminator": {
          "mapping": {
           "assumption": "#/components/schemas/Assumption",
           "criterion": "#/components/schemas/Criterion",
           "decision": "#/components/schemas/DecisionItem",
           "evidence": "#/components/schemas/Evidence",
           "feedback": "#/components/schemas/Feedback",
           "task": "#/components/schemas/Task"
          },
          "propertyName": "type"
         },
         "oneOf": [
          {
           "$ref": "#/components/schemas/Assumption"
          },
          {
           "$ref": "#/components/schemas/Criterion"
          },
          {
           "$ref": "#/components/schemas/Task"
          },
          {
           "$ref": "#/components/schemas/Evidence"
          },
          {
           "$ref": "#/components/schemas/DecisionItem"
          },
          {
           "$ref": "#/components/schemas/Feedback"
          }
         ],
         "title": "Response Update Item Api Projects  Project Id  Items  Item Id  Patch"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Item",
    "tags": [
     "items"
    ]
   }
  },
...
 }
}